+++ R/vroom_write.R
@@ -46,10 +46,6 @@ vroom_write <- function(x, file, delim = '\t', eol = "\n", na = "NA", col_names 
     )
   }
 
-  # If there are no columns in the data frame, just return
-  if (NCOL(x) == 0) {
-    return(invisible(x))
-  }
 
   quote <- match.arg(quote)
   escape <- match.arg(escape)
@@ -58,6 +54,14 @@ vroom_write <- function(x, file, delim = '\t', eol = "\n", na = "NA", col_names 
 
   # Standardise path returns a list, but we will only ever have 1 output file.
   file <- standardise_one_path(file, write = TRUE)
+
+  # If there are no columns in the data frame, just create an empty file and return
+  if (NCOL(x) == 0) {
+    if (!inherits(file, "connection")) {
+      file.create(file)
+    }
+    return(invisible(x))
+  }
 
   # We need to convert any altrep vectors to normal vectors otherwise we can't fill the
   # write buffers from other threads.