+++ NEWS.md
@@ -1,3 +1,5 @@
+# vroom (development version)
+
 # vroom 1.5.3
 
 * `vroom_write()` now creates an empty file when given a input without any columns (https://github.com/tidyverse/readr/issues/1234)