+++ NEWS.md
@@ -1,12 +1,13 @@
 # vroom (development version)
 
+* `vroom(col_types=)` now accepts column type names like those accepted by utils::read.table. e.g.
+    vroom::vroom(col_types = list(a = "integer", b = "double", c = "skip"))
+
 * `vroom()` now respects the `quote` parameter properly in the first two lines of the file (https://github.com/tidyverse/readr/issues/1262)
 
 * `vroom_write()` now always correctly writes its output including column names in UTF-8 (https://github.com/tidyverse/readr/issues/1242)
 
 * `vroom_write()` now creates an empty file when given a input without any columns (https://github.com/tidyverse/readr/issues/1234)
-
-* `vroom(col_types=)` now accepts column type names like those accepted by utils::read.table.
 
 # vroom 1.5.3